Refactored layout of modelstate class again
